requirements.txt
streamlit==1.39.0
pydub==0.25.1
ffmpeg-python==0.2.0

faster-whisper==1.0.3
ctranslate2==4.4.0

notion-client==2.2.1

google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
